backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.9
pillow>=10.0.0
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.11.0
python-magic>=0.4.27
aiofiles>=23.2.1
requests>=2.31.0
huggingface-hub>=0.17.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0